Implement auth functionality and fix related issues to align with GraphQL, JWT decoding issue and update & relocated signup/signin pages
