+++ Fitzys_Fades/client/src/utils/mutations.js
@@ -13,3 +13,11 @@ mutation CreateUser($userInput: UserInput!) {
     }
   }
   `;
+
+  export const LOGIN_USER = gql`
+  mutation LoginUser($email: String!, $password: String!) {
+    loginUser(email: $email, password: $password) {
+      token
+    }
+  }
+`;
