+++ Fitzys_Fades/client/src/utils/queries.js
@@ -1,0 +1,11 @@
+import { gql } from '@apollo/client';
+
+export const GET_USER_INFO = gql`
+  query GetUser($userId: ID!) {
+    user(id: $userId) {
+      id
+      name
+      email
+    }
+  }
+`;
